Move quoteIfNeeded to sql.catalyst.util.
